STM32: Refactor lp_ticker.c + rtc_api.c + sleep.c + rtc_api_hal.h files
