+++ targets/TARGET_STM/TARGET_STM32F3/common_objects.h
@@ -116,5 +116,8 @@ struct i2c_s {
 }
 #endif
 
+/* STM32F3 HAL doesn't provide this API called in rtc_api.c */
+#define __HAL_RCC_RTC_CLKPRESCALER(__RTCCLKSource__)
+
 #endif
