+++ targets/TARGET_STM/lp_ticker.c
@@ -49,13 +49,13 @@ void lp_ticker_init(void)
 
 uint32_t lp_ticker_read(void)
 {
-    uint32_t usecs;
-    time_t time;
+    uint32_t usecs = 0;
+    time_t time = 0;
 
     lp_ticker_init();
 
     do {
-    time = rtc_read();
+      time = rtc_read();
       usecs = rtc_read_subseconds();
     } while (time != rtc_read());
 
@@ -68,7 +68,7 @@ void lp_ticker_set_interrupt(timestamp_t timestamp)
 
     delta = timestamp - lp_ticker_read();
     rtc_set_wake_up_timer(delta);
-    }
+}
 
 void lp_ticker_disable_interrupt(void)
 {
