+++ targets/TARGET_STM/rtc_api_hal.h
