+++ targets/TARGET_STM/sleep.c
@@ -1,6 +1,6 @@
 /* mbed Microcontroller Library
  *******************************************************************************
- * Copyright (c) 2014, STMicroelectronics
+ * Copyright (c) 2016, STMicroelectronics
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -35,7 +35,8 @@
 #include "cmsis.h"
 
 
-void sleep(void) {
+void sleep(void)
+{
     // Stop HAL systick
     HAL_SuspendTick();
     // Request to enter SLEEP mode
@@ -44,15 +45,24 @@ void sleep(void) {
     HAL_ResumeTick();
 }
 
-
 void deepsleep(void)
 {
+    // Stop HAL systick
+    HAL_SuspendTick();
+
     // Request to enter STOP mode with regulator in low power mode
+#if TARGET_STM32L4
+    HAL_PWREx_EnterSTOP2Mode(PWR_STOPENTRY_WFI);
+#else /* TARGET_STM32L4 */
     HAL_PWR_EnterSTOPMode(PWR_LOWPOWERREGULATOR_ON, PWR_STOPENTRY_WFI);
+#endif /* TARGET_STM32L4 */
+
+    // Restart HAL systick
+    HAL_ResumeTick();
 
     // After wake-up from STOP reconfigure the PLL
     SetSysClock();
-    
+
 #if DEVICE_LOWPOWERTIMER
     rtc_synchronize();
 #endif
